cpp in done done

+++ cipipi09/ex01/RPN.hpp
@@ -19,7 +19,7 @@
 class RPN
 {
 	private:
-        std::stack<int> stack;
+        std::stack<double> stack;
 		
 	public:
 		RPN();
